requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.6
django-crispy-forms==1.5.2
django-haystack==2.4.1
django-toolbelt==0.0.1
djangorestframework==3.3.3
elasticsearch==1.7.0
gunicorn==19.4.5
Pillow==3.1.
pyelasticsearch==1.4
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.0.11
static3==0.6.1
urllib3==1.14
whitenoise==2.0.6
Whoosh==2.7.0
